kernel: inet stack clean up, change memory allocations, add error checks

+++ moose/include/drivers/rtl8139.h
@@ -6,4 +6,4 @@
 #define RTL8139_DEVICE_ID 0x8139
 
 int init_rtl8139(u8 *mac_addr);
-void rtl8139_send(void *frame, u16 size);
+void rtl8139_send(void *frame, size_t size);
